fixing some issues caused by the roscpp bug that adds '...' at the end of parameter values
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@23034 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 67eb58a32dbaf15f14d45efb8adf24b726544bb2

+++ motion_planning/planning_environment/src/tools/clear_known_objects.cpp
@@ -273,7 +273,7 @@ private:
 			data_out.channels[j].values.push_back(cloud->channels[j].values[i]);
 		}
 
-	    ROS_DEBUG("Published filtered cloud (%d points out of %d)", data_out.points.size(), cloud->points.size());
+	    ROS_DEBUG("Published filtered cloud (%d points out of %d)", (int)data_out.points.size(), (int)cloud->points.size());
 	    cloudPublisher_.publish(data_out);
 	}
 	else
@@ -378,7 +378,6 @@ private:
    
 int main(int argc, char **argv)
 {
-  // Took out anonymous - bmm
     ros::init(argc, argv, "clear_known_objects");
 
     ClearKnownObjects cko;
